Cleanup + pheremones and cross-creature awareness
Creatures can now see and lay down pheremones, which allows them to, in a convoluted way, communicate with eachother and create some form of cross-creature communication. The next update to the project will allow for more in depth and detailed communication.

On top of all of this, the creatures can now "sense" the creature closest to them, allowing for spatial awareness across these creatures.